+++ Assets/Scripts/Fighting/HealthSystem.cs
@@ -5,10 +5,10 @@ public class HealthSystem : MonoBehaviour, IDamagable
 {
     private const int MinHealthPoints = 0;
 
-    [field: SerializeField] public int MaxHealthPoints { get; private set; }
-    [field: SerializeField] public int HealthPoints { get; private set; }
+    [field: SerializeField] public float MaxHealthPoints { get; private set; }
+    [field: SerializeField] public float HealthPoints { get; private set; }
 
-    public event Action<int> HealthChanged;
+    public event Action<float> HealthChanged;
     public event Action TookDamage;
 
     private void OnValidate()
@@ -17,7 +17,7 @@ public class HealthSystem : MonoBehaviour, IDamagable
             HealthPoints = MaxHealthPoints;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         if (damage < 0)
             throw new ArgumentOutOfRangeException(nameof(damage), "Damage cannot be negative.");
@@ -27,7 +27,7 @@ public class HealthSystem : MonoBehaviour, IDamagable
         TookDamage?.Invoke();
     }
 
-    public void Heal(int heal)
+    public void Heal(float heal)
     {
         if (heal < 0)
             throw new ArgumentOutOfRangeException(nameof(heal), "Healing cannot be negative.");
